js: Update prettier, eslint, prettier-eslint, and prettier-eslint-cli
Following the eagerly awaited resolution of
  https://github.com/prettier/prettier-eslint-cli/issues/304
.

And adjust our ESLint config so that we can do this with no new
ESLint errors or suppressions. After this, we'll un-ignore the rules
under "New rules we want", one by one, which will mean fixing the
code that doesn't yet follow each rule.

Greg says, about the formatting changes:
  https://github.com/zulip/zulip-mobile/pull/5393#pullrequestreview-986977782

> The formatting changes from the Prettier upgrade look fine. Some
> code gets nicer (like `foo.bar().baz()` going on one line), some
> gets less nice (like a lot of stuff getting an extra level of
> indentation). I'm not sure I like it better on net; but Prettier
> is designed as a take-it-or-leave-it package, and I don't have a
> better alternative nor want to be stuck on an old version. So
> :shrug:

The added @babel/* deps are to satisfy peer-dependency requirements.

The resolutions line gets us
  https://github.com/prettier/prettier-eslint/pull/749
and can be removed when prettier-eslint-cli bumps its
prettier-eslint version.

And run

  $ tools/tsflower unpack
  $ tools/tsflower pack

to update TsFlower's output for the new Prettier version and rebase
the patches atop that. I see some churn in the patches where it
looks like a Git version is printed at the bottom, and mine differs
from Greg's, hmm.