users state [nfc]: Pass realm_user update events through more directly
Following the generic "server event" pattern Greg started in
979d28358. We still convert avatar_url to our special AvatarURL form
at the edge, in eventToAction, for crunchy shell:
  https://github.com/zulip/zulip-mobile/blob/master/docs/architecture/crunchy-shell.md

Also, give a Flow type to the event itself as we receive it from the
server: RealmUserUpdateEventRaw. We should keep this up-to-date with
the server API documentation. (We also have RealmUserUpdateEvent,
for the slightly processed form that has AvatarURL.)

We still keep the behavior of only allowing avatar_url and role
changes to apply to the state [1]. But now, that code is in
usersReducer instead of eventToAction. For the reducer tests that
simulate actions changing *other* user data, that means they'd now
fail. So, skip them, with a plan to re-enable.

[1] Because we've been wary of accepting server data that we haven't
    vetted; see 729655187. The API doc seems to have been improved
    since then, and now we've written down a type for it. So now we
    can proceed more confidently, which is nice!