EditStreamScreen [nfc]: Stop changing initialValues sent to EditStreamCard
This is NFC because EditStreamCard has already been using a snapshot
of initialValues that it makes on its first render. But since the
prop is named `initialValues`, we shouldn't be passing something
that changes to be different from the initial values.

This will help with an upcoming change; see Greg's comment at
  https://github.com/zulip/zulip-mobile/pull/5389#discussion_r881187606