gitignore: Ignore node_modules only where expected, at root
If one of these ever appears in an unexpected place, it could
cause some very confused debugging.  Best to avoid.

A draft of the TsFlower branch did result in a couple of these.
They were pretty benign, but caused me to notice this rule's breadth.

Also drop the entry for `npm-debug.log`, which doesn't appear in
our workflows.