tools/tsflower [nfc]: Dedupe shared logic between packages
While here, clean up to more consistently use `rootdir`.  Though
that should still be NFC, because we cd there at the outset anyway.

Also, since we're now writing a function that's supposed to work on
any package in node_modules: instead of hardcoding where in the
package's tree the type definitions are found, get it from the
package.json.