nav types: Fix RouteProp, using $Exact; cut redundant-intersection hack
Since the recent commit 067a450a4 "types: Switch to TsFlower for
react-navigation types", `props.route.params` has been ending up
with a type of `empty`, in our screens where we consume that.
Yikes!  That means it's treated as a magic shapeshifter value --
the uses of it are effectively not type-checked.

Because that state of affairs only developed recently, no genuine
type errors snuck in past it.  It did have the effect of suppressing
an error we'd have otherwise gotten at the initialParams uses in
SharingScreen, which was due to the oddity in the library type
definition fixed in the preceding commit.

The type of `props.route` on each screen comes from `RouteProp`,
so that's where we need to make a fix.

Applying `$Exact` to the inexact `Route` type, before we go and
spread it into this exact object type, fixes the issue.

After doing that, we can also remove the redundant intersection we'd
had here; those puzzling errors no longer recur.