nav: Switch from NavigationService to props.navigation
React Navigation upstream has since the 5.x release been
discouraging the use of the strategy we call NavigationService.
See discussion at 140c28cc1 and:
  https://github.com/zulip/zulip-mobile/issues/4417
Instead, the recommendation is to use the `navigation` prop where
available, and failing that the `useNavigation` hook.

In this commit, we make that change for all sites where we were
using NavigationService in a function component that is a screen
in a navigator, and gets its own navigation prop.