nav types: Fix SettingsScreen's navigation-prop type
We moved this screen a while ago from being one of the main tabs to
being a screen reached by stack navigation.  Since then, this type
has been invalid because 'settings' is not among the route keys in
the "main tabs" navigator.

Flow hasn't been giving us an error about it, because we haven't had
any actual value expressions around that have this type -- we never
look at the `navigation` prop in this component.  We do so in the
next commit, which caused me to notice this error.  Fix it.