tools/tsflower: Make patches reproducible across diff.interHunkContext
This was causing a small bit of churn (not as much as the
Git-version trailer) when going between patches generated on
my machine and on Chris's.

I think the value I have in my local config (and reflected
unintentionally in the first versions of these patches) tends to make
diffs a bit easier to read than the default, so go back to that.