deps: Add expo-cli at latest, in dev dependencies
We'll use this soon when we upgrade `expo` to 44.

I assume, all else being equal, it's better to put this in
devDependencies than install it globally, since we want commands to
have consistent results across environments. (The guide we're about
to follow, to upgrade to Expo 44, would have us install it
globally.) It comes with a ton of new dependencies, though, and that
might be a reason to not have it in the project.

And add .expo/ to our gitignore. It gets created when you run
`expo upgrade`, containing a README that says it shouldn't be
committed. And an Expo template app has it in the gitignore:
  https://github.com/expo/expo/blob/202e7517/templates/expo-template-bare-minimum/gitignore#L53

And run `yarn yarn-deduplicate && yarn`, as prompted by
`tools/test deps`.