deps: Upgrade to Expo SDK 44
Following the "bare workflow" instructions for upgrading:
  https://blog.expo.dev/expo-sdk-44-4c4b8306584a#8ff0

When I ran `expo upgrade`, it went through a phase it described as
"Updating packages to compatible versions (where known)." This
changed a lot of version ranges, even of non-Expo things like
react-native-reanimated. I kept some of these and ignored others,
and I plan to elaborate on these choices in the GitHub PR thread. I
didn't find more info in the CLI's docs about how it chooses these
upgrades. I believe the code is
  https://github.com/expo/expo-cli/blob/expo-cli%405.4.12/packages/expo-cli/src/commands/info/upgradeAsync.ts#L541-L597
and that
  https://github.com/expo/expo/blob/sdk-44/packages/expo/bundledNativeModules.json
gets consulted as part of that.

In the "Update your `App.Delegate.m` according to this diff" step, I
applied the substance of the change, with a few formatting
differences to minimize our diff, on top of some changes I made for
the expo-screen-orientation upgrade. That upgrade trimmed out some
setup requirements for the package:
  https://github.com/expo/expo/blob/main/packages/expo-screen-orientation/CHANGELOG.md#410--2021-12-03