flow: Use a Flow enum for the first time! (CreateWebPublicStreamPolicy)
This is nice and simple, particularly in the definition.

We remove the `default` case in the two `switch`es on a
CreateWebPublicStreamPolicy value, and in particular that means
removing the `ensureUnreachable`s there. But its job will still be
done: Flow still tracks whether the cases are exhaustive and will
complain if not.