SwitchRow: Use container minHeight for height
The actual touch target on SwitchRow is the switch. That's a native
platform UI component that we can't resize, and we leave it to the
platform to render as it wants. The inspector showed 31px height for
the switch on iOS in my testing.

Might as well still give the container a minHeight of 48 for the
reason given in the comment.