nav layout: Have main-tabs screens consume top inset individually
Instead of consuming it in their shared parent component,
MainTabsScreen.

Now, when we add app bars on each of these screens -- which, oops,
we did for SubscriptionsScreen in 9d11e787f without prepping with
this first -- the app-bar component can properly consume the top
inset, covering it with its own appropriate, contentless surface.

In particular, to do so "properly", the top inset must not be
double-consumed by MainTabsScreen; fix that.

So, this fixes a regression where we had twice as much top-padding
as we needed to account for the top inset in the "Streams" screen
(SubscriptionsScreen).