emoji [nfc]: Have displayCharacterForUnicodeEmojiCode parse emoji codes
Instead of doing it in the construction of codeToEmojiMap.

We're about to have displayCharacterForUnicodeEmojiCode consume
state.realm.serverEmojiData, while keeping the codepath that
consults our hard-coded data as a fallback.

Since state.realm.serverEmojiData only gives us emoji codes in their
unparsed form, we'll need to call parseUnicodeEmojiCode on those
emoji codes. It'll be neater to do the same in the fallback path
too, so the two paths are parallel in that respect.