url: Consistently parse URLs for isUrlOnRealm (CVE-2022-35962)
The ad-hoc parsing we'd been doing here had the consequence that it
could return true for some URL strings that in reality resolve to
URLs that aren't on the Zulip realm.

If another (authenticated) user sends a message containing an image
link crafted to trigger this bug, and the receiving user taps on the
image in order to expand it in the lightbox, this could cause the
user's login credentials to be disclosed.  This is CVE-2022-35962.

To fix the issue, use `new URL` to parse the URL, the same way we're
going to do at each of its two call sites.

This still isn't a great structure, because it'd be easy for a call
site to change (or a new one to be added) so that the URL gets used
in a way other than by passing to our `new URL`, or with a different
base URL.  Add a TODO to refactor accordingly, and mark the function
as deprecated to discourage new callers.