settings screen [nfc]: Simplify comparing *Row layouts on this screen
Now, there's less clicking required to check, by reading code, that
the three components NestedNavRow, SwitchRow, and
InputRowRadioButtons will look reasonably uniform. That's nice when
they're all seen together, as they are when they make a list of
settings on the settings screen (with the last of these being a
newcomer there; see recent commits).

This also continues our progress toward removing the constants in
miscStyles.