react: Convert class components to function components, where trivial
None of these have any custom instance methods, or lifecycle methods
other than `render`, so they're nice and boring and easy.

This doesn't mean all the rest of our React components will be hard
to convert. These were just the (very) low-hanging fruit. :)

As usual, passing -b to a git-log command is helpful to reduce noise
when viewing a class-to-function-component conversion.