messagesActions [nfc]: Pull out adjustNarrowForMoves from doNarrowNearLink
This gives a bit of separation between the tasks of "figure out where
we should narrow" -- which can need a network request to get some
data, but should have no further side effects -- and "OK, narrow
there", which is all about side effects.

As a result it simplifies all the early-return paths, and deduplicates
the call to `doNarrow`.