msglist: Convert to a Hooks-based function component
Most of this is a routine class-to-Hooks conversion.  The
interesting part is the lifecycle method `shouldComponentUpdate`;
the part where it tracks changes to props turns into an effect,
and the part where it blocks (or tries to) re-renders turns into
forcibly memoizing the bulk of the old render method using a ref.

As a bonus, this fixes a longstanding latent bug: React could
have at any time chosen to go ahead and re-render the component
even though we have a `shouldComponentUpdate` that always returns
false.  (The React docs stress that there's no guarantee on that
score, that the method exists only for performance optimization.)

That would cause us to recompute `html`, potentially taking some
time.  And if the resulting value had changed since the initial
render, it'd cause the state in the WebView, including scroll
position, to reset.

With React Hooks, we can use refs to ensure that we never compute
`html` a second time, and never change its value, just as we intend.
(With a bit of future-proofing to ensure we continue displaying
something coherent if the theme changes, which #5533 will enable.)